refactor: move diagnostics logic to shared module
